Set up expanded state for StudentCard
This commit adds an `expand` boolean state to the <StudetnCard>
component. When true card will display all of the students grades.
When false grades will not be shown.

future commits will implement a button to toggle the state between
true and false

For the moment, **there are no user facing changes** in this commit.